Rationalise carrying forward custom metadata
CGI.pm returns multiple paramters with the same name (such as those in
our "Custom fields" metadata) packed with a null byte (\0) separator.
For some reason, I suspect an oversight on my part, we were splitting
this into an array and joining them back together with a tab separator
in order to carry them forward. This is entirely unecessary. We still
need to split them occassionally, for example when doing a non-JS delete
of a custom field, but we always join them back up with \0 and retain
the original name