Make "borrowers" stage more resilient
We were hardcoding the properties to carry forward. This is super prone
to getting left behind so we now iterate what we've been passed and just
pass it forward.

We have to jump through some hoops to pass forward custom_{key,value}
as, for some reason I don't understand, TT is converting the null byte
separator with an undefined charater, which isn't what we want, so we
have to manually split that out into it's constituent properties